CLI-65: Refactor env helpers
