+++ test/ConfigManagementHelper.js
@@ -20,10 +20,36 @@ const clonedeep = require('lodash.clonedeep');
 const moment = require('moment');
 
 const ApiService = require('./ApiService');
+const { BackendCollectionPermission, CollectionHook } = require('./../lib/Constants');
 const TestsHelper = require('./TestsHelper');
-const { writeJSON } = require('./../lib/Utils');
+const { isEmpty, writeJSON } = require('./../lib/Utils');
 
 let ConfigManagementHelper = {};
+
+function passConfigFileToCli(cmd, configContent, filePath, done) {
+  async.series([
+    (next) => {
+      writeJSON(filePath, configContent, next);
+    },
+    (next) => {
+      TestsHelper.execCmdWoMocks(cmd, null, (err, data) => {
+        if (err) {
+          return next(err);
+        }
+
+        const parsedData = JSON.parse(data);
+        const id = parsedData.result.id;
+        next(null, id);
+      });
+    }
+  ], (err, results) => {
+    if (err) {
+      return done(err);
+    }
+
+    done(null, results.pop());
+  });
+}
 
 const env = {};
 env.buildSettings = function buildSettings(options) {
@@ -75,6 +101,261 @@ env.buildValidInternalCollectionsList = function buildValidInternalCollectionsLi
   return result;
 };
 
+env.createFromConfig = function createEnvFromConfig(envName, env, appName, done) {
+  const fileName = `${TestsHelper.randomStrings.plainString(10)}.json`;
+  const filePath = path.join(TestsHelper.ConfigFilesDir, fileName);
+  const cmd = `env create ${envName} ${filePath} --app ${appName} --output json`;
+  passConfigFileToCli(cmd, env, filePath, done);
+};
+
+env.assertEnvOnly = function assertEnvOnly(envFromConfig, envId, envName, done) {
+  ApiService.envs.get(envId, (err, actualEnv) => {
+    if (err) {
+      return done(err);
+    }
+
+    expect(actualEnv.name).to.equal(envName);
+
+    const envSettings = envFromConfig.settings;
+    if (envSettings && envSettings.emailVerification) {
+      expect(actualEnv.emailVerification).to.deep.equal(envSettings.emailVerification);
+    } else {
+      expect(actualEnv).to.not.have.property('emailVerification');
+    }
+
+    let expectedApiVersion = 3;
+    if (envSettings && envSettings.apiVersion) {
+      expectedApiVersion = envSettings.apiVersion;
+    }
+
+    expect(actualEnv.apiVersion).to.equal(expectedApiVersion);
+
+    done();
+  });
+};
+
+env.buildExpectedPermissionsPerColl = function buildExpectedPermissionsPerColl(collList, rolesNameIdPairs) {
+  const expectedPermissionsPerColl = {};
+
+  collList.forEach((coll) => {
+    let expectedPermissions;
+    const collPermissionsInConfig = coll.permissions;
+    if (typeof collPermissionsInConfig === 'string') {
+      expectedPermissions = BackendCollectionPermission[collPermissionsInConfig];
+    } else {
+      expectedPermissions = {};
+      const rolesNames = Object.keys(collPermissionsInConfig);
+      rolesNames.forEach((roleName) => {
+        const permissionsPerRole = Object.keys(collPermissionsInConfig[roleName]);
+        permissionsPerRole.forEach((permission) => { // e.g. create, update
+          if (!expectedPermissions[permission]) {
+            expectedPermissions[permission] = [];
+          }
+
+          expectedPermissions[permission].push({
+            roleId: rolesNameIdPairs[roleName] || 'all-users',
+            type: collPermissionsInConfig[roleName][permission]
+          });
+        });
+      });
+    }
+
+    expectedPermissionsPerColl[coll.name] = expectedPermissions;
+  });
+
+  return expectedPermissionsPerColl;
+};
+
+env.assertCollections = function assertCollections(envId, collList, expCollCount, rolesNameIdPairs, done) {
+  ApiService.colls.get(envId, null, (err, actualColls) => {
+    if (err) {
+      return done(err);
+    }
+
+    expect(expCollCount).to.equal(actualColls.length);
+
+    const expectedPermissionsPerColl = env.buildExpectedPermissionsPerColl(collList, rolesNameIdPairs);
+    const collsFromConfigCount = collList.length;
+
+    for (let i = 0; i < collsFromConfigCount; i += 1) {
+      const expColl = collList[i];
+      const actualColl = actualColls.find(x => x.name === expColl.name);
+      if (!actualColl) {
+        return done(new Error(`Failed to find coll with name '${expColl.name}'.`));
+      }
+
+      const expPermissions = expectedPermissionsPerColl[actualColl.name];
+      expect(actualColl.permissions).to.deep.equal(expPermissions);
+
+      if (expColl.type === 'internal') {
+        expect(actualColl.dataLink).to.be.null;
+      } else {
+        expect(actualColl.dataLink).to.exist;
+        expect(actualColl.dataLink.serviceObjectName).to.equal(expColl.serviceObject);
+      }
+    }
+
+    done();
+  });
+};
+
+env.assertCollHooksPerColl = function assertCollHooksPerColl(envId, collName, collHooks, done) {
+  const collHooksNames = Object.keys(collHooks);
+
+  async.eachSeries(
+    collHooksNames,
+    (currentHook, next) => {
+      const backendHookName = CollectionHook[currentHook];
+      ApiService.businessLogic.collHooks.get(envId, collName, backendHookName, (err, actualHook) => {
+        if (err) {
+          return next(err);
+        }
+
+        const expectedHook = collHooks[currentHook];
+        if (expectedHook.type === 'internal') {
+          expect(actualHook.host).to.be.null;
+        } else {
+          expect(actualHook.host).to.not.be.null;
+        }
+
+        const defaultCode = `function ${currentHook}(request, response, modules) {\n  response.continue();\n}`;
+        const expectedCode = expectedHook.code || defaultCode;
+        expect(expectedCode).to.equal(actualHook.code);
+
+        if (expectedHook.type === 'internal') {
+          expect(actualHook.host).to.be.null;
+        } else {
+          expect(actualHook.host).to.not.be.null;
+          expect(actualHook.sdkHandlerName).to.equal(expectedHook.handlerName);
+        }
+
+        next();
+      });
+    },
+    done
+  );
+};
+
+env.assertAllCollHooks = function assertAllCollHooks(envId, configHooks, done) {
+  const collNames = Object.keys(configHooks);
+
+  async.eachSeries(
+    collNames,
+    (currentCollName, next) => {
+      env.assertCollHooksPerColl(envId, currentCollName, configHooks[currentCollName], next);
+    },
+    done
+  );
+};
+
+env.assertEndpoints = function assertEndpoints(envId, configEndpoints, done) {
+  const endpointNames = Object.keys(configEndpoints);
+
+  async.eachSeries(
+    endpointNames,
+    (currentEndpointName, next) => {
+      ApiService.businessLogic.endpoints.get(envId, currentEndpointName, (err, actual) => {
+        if (err) {
+          return done(err);
+        }
+
+        const expected = configEndpoints[currentEndpointName];
+        if (expected.type === 'internal') {
+          expect(actual.host).to.be.null;
+        } else {
+          expect(actual.host).to.not.be.null;
+        }
+
+        const defaultCode = 'function onRequest(request, response, modules) {\n  response.continue();\n}';
+        const expectedCode = expected.code || defaultCode;
+        expect(expectedCode).to.equal(actual.code);
+
+        if (expected.schedule) { // due to api peculiarities
+          if (!expected.schedule.interval) {
+            expect(expected.schedule.start).to.equal(actual.schedule);
+          } else {
+            expect(expected.schedule.start).to.equal(actual.schedule.start);
+            expect(expected.schedule.interval).to.equal(actual.schedule.interval);
+          }
+        } else {
+          expect(actual.schedule).to.be.null;
+        }
+
+        if (expected.type === 'internal') {
+          expect(actual.host).to.be.null;
+        } else {
+          expect(actual.host).to.not.be.null;
+          expect(actual.sdkHandlerName).to.equal(expected.handlerName);
+        }
+
+        next();
+      });
+    },
+    done
+  );
+};
+
+env.assertGroups = function assertGroups(envId, configGroups, done) {
+  const groupNames = Object.keys(configGroups);
+
+  async.eachSeries(
+    groupNames,
+    (currentGroupName, next) => {
+      ApiService.groups.get(envId, currentGroupName, (err, actual) => {
+        if (err) {
+          return next(err);
+        }
+
+        expect(actual).to.have.property('_acl');
+        expect(actual).to.have.property('_kmd');
+
+        const expected = configGroups[currentGroupName];
+        if (isEmpty(expected)) {
+          return setImmediate(next);
+        }
+
+        expect(expected.name).to.equal(actual.name);
+        expect(expected.description).to.equal(actual.description);
+
+        if (!isEmpty(expected.groups)) {
+          expect(expected.groups.length).to.equal(actual.groups.length);
+
+          for (const expGroup of expected.groups) {
+            const foundGroup = actual.groups.find(x => x._id === expGroup);
+            if (!foundGroup) {
+              return next(new Error(`Could not find group with identifier '${expGroup}'.`));
+            }
+          }
+        }
+
+        next();
+      });
+    },
+    done
+  );
+};
+
+env.assertPushSettings = function assertPushSettings(envId, configPushSettings, done) {
+  ApiService.push.get(envId, (err, actual) => {
+    if (err) {
+      return done(err);
+    }
+
+    if (!isEmpty(configPushSettings.android)) {
+      expect(configPushSettings.android.senderId).to.equal(actual.android.projectId);
+      expect(configPushSettings.android.apiKey).to.equal(actual.android.apiKey);
+    } else {
+      expect(actual.android).to.be.false;
+    }
+
+    if (isEmpty(configPushSettings.ios)) {
+      expect(actual.ios).to.be.false;
+    }
+
+    done();
+  });
+};
+
 function buildConfigEntityFromList(list) {
   const result = {};
 
@@ -86,31 +367,6 @@ function buildConfigEntityFromList(list) {
   });
 
   return result;
-}
-
-function passConfigFileToCli(cmd, configContent, filePath, done) {
-  async.series([
-    (next) => {
-      writeJSON(filePath, configContent, next);
-    },
-    (next) => {
-      TestsHelper.execCmdWoMocks(cmd, null, (err, data) => {
-        if (err) {
-          return next(err);
-        }
-
-        const parsedData = JSON.parse(data);
-        const id = parsedData.result.id;
-        next(null, id);
-      });
-    }
-  ], (err, results) => {
-    if (err) {
-      return done(err);
-    }
-
-    done(null, results.pop());
-  });
 }
 
 const app = {};
